Patch 0913
・影狼、朧スキル「明鏡止水」のSP回復不具合を修正 (src/map/status.c)
	Thanks かんなぎ さん
・拡張スーパーノービス、影狼、朧の秒間最大攻撃回数が固定になる不具合を修正 (src/map/status.c)
	Thanks barusu2 さん、rapis さん
・二刀流時の右手・左手修練のダメージ計算式を修正 (src/map/battle.c)
	Thanks dragontail789 さん
・add_damage_class_count、add_damage_class_count_、add_magic_damage_class_countはintである必要がないのでshortに変更
 (src/map/map.h)

+++ src/map/map.h
@@ -530,7 +530,7 @@ struct map_session_data {
 	int critical_rate,hit_rate,flee_rate,flee2_rate,def_rate,def2_rate,mdef_rate,mdef2_rate;
 	int def_ratio_atk_ele,def_ratio_atk_race,def_ratio_atk_enemy;
 	int def_ratio_atk_ele_,def_ratio_atk_race_,def_ratio_atk_enemy_;
-	int add_damage_class_count,add_damage_class_count_,add_magic_damage_class_count;
+	short add_damage_class_count,add_damage_class_count_,add_magic_damage_class_count;
 	short add_damage_classid[10],add_damage_classid_[10],add_magic_damage_classid[10];
 	int add_damage_classrate[10],add_damage_classrate_[10],add_magic_damage_classrate[10];
 	short add_def_class_count,add_mdef_class_count;
